requirements.txt
PyPDF2
google-generativeai
python-dotenv
requests
tqdm